Create a HAL classFirst step - abstract motor control to HAL
