+++ src/config.h
@@ -4,11 +4,6 @@
 #include <unordered_map>
 using namespace std;
 
-// M_PI
-#ifndef M_PI
-    #define M_PI 3.14159265358979323846
-#endif
-
 // Pins
 #ifdef ARDUINO_TEENSY31
 #include "config/teensy.h"
@@ -18,7 +13,6 @@ using namespace std;
 #include "config/blackpill.h"
 #endif
 
-
 // i2c Addresses
 #define TCA_ADDR 0x70
 #define DISPLAY_ADDR 0x3c
@@ -27,12 +21,11 @@ using namespace std;
 #define ADC_ADDR 0x48
 typedef const std::unordered_map<uint8_t, String> HexIntMap;
 HexIntMap I2C_ADDRESS_NAMES = {
-   {TCA_ADDR, "Multiplexer"},
-   {IMU_ADDR, "IMU"},
-   {TOF_ADDR, "Distance sensor"},
-   {DISPLAY_ADDR, "OLED display"},
-   {ADC_ADDR, "Minesweeper ADC"}
-};
+    {TCA_ADDR, "Multiplexer"},
+    {IMU_ADDR, "IMU"},
+    {TOF_ADDR, "Distance sensor"},
+    {DISPLAY_ADDR, "OLED display"},
+    {ADC_ADDR, "Minesweeper ADC"}};
 // hardware counts
 #define NUM_ENCODERS 6
 
@@ -42,4 +35,4 @@ HexIntMap I2C_ADDRESS_NAMES = {
 // Comms constants
 #define MAX_MISSED_MOTOR_MESSAGES 10
 
-#endif //_CONFIG__H_
+#endif  //_CONFIG__H_
