+++ src/utils.cpp
@@ -1,0 +1,24 @@
+#include "utils.h"
+
+float minMagnitude(float x, float y, float z) {
+    //function to find the smallest (closest to zero) value
+    // specifically, find the motor that is spinning slowest
+    // which is assumed to be the most representative of robot speed
+    float currentMin = x;
+    float currentMinMagnitude = abs(x);
+    if (abs(y) < currentMinMagnitude) {
+        currentMin = y;
+        currentMinMagnitude = abs(y);
+    }
+    if (abs(z) < currentMinMagnitude) {
+        currentMin = z;
+    }
+    return currentMin;
+}
+
+float wrapTwoPi(float angle) {
+    //wraps an angle to stay within +/-pi
+    while (angle > M_PI) angle -= TWO_PI;
+    while (angle < -M_PI) angle += TWO_PI;
+    return angle;
+}
